further refactor controllers with UtilController class
